Introduction + descriptions
I tried to add the themis package to balance the number of absences and presence but the code just doesn't seem to work. I put my draft in and send a question on plaza, if you guys can figure it out, it would be awesome. Refer to 5.7.2 in the textbook.